+++ packages/moonstone/CHANGELOG.md
@@ -6,8 +6,20 @@ The following is a curated list of changes in the Enact moonstone module, newest
 
 ### Fixed
 
-- `moonstone/ContextualPopupDecorator` layout in large text mode in RTL locales
+- `moonstone/Scroller` to restore focus properly when pressing page up after holding 5-way down
 - `moonstone/VirtualList.VirtualList` dynamically extended item scrolling into view properly
+
+## [3.0.0] - 2019-09-03
+
+### Fixed
+
+- `moonstone/ContextualPopupDecorator` layout in large text mode in RTL locales
+- `moonstone/Dropdown` performance when using many options
+- `moonstone/ProgressBar` fill color when `highlighted` is set
+- `moonstone/Scroller` to correctly handle horizontally scrolling focused elements into view when using a `direction` value of `'both'`
+- `moonstone/Skinnable` TypeScript signature
+- `moonstone/Slider` progress bar fill color when focused with `noFill` set
+- `moonstone/VirtualList.VirtualGridList` and `moonstone/VirtualList.VirtualList` to render the first item properly when the `dataSize` prop is updated and the function as a parameter of the `cbScrollTo` prop is called
 
 ## [3.0.0-rc.4] - 2019-08-22